Разделяет основной скрипт на модули data и utils
