--- js/utils/check-string-length.js
@@ -1,5 +1,0 @@
-function checkStringLength (string, length) {
-  return string.length <= length;
-}
-
-checkStringLength('Hello World!', 140);
